ADM-623:[frontend]feat: fix the issue of branch selector, when click away, popper will miss.
