+++ frontend/src/components/Metrics/MetricsStep/index.tsx
@@ -4,13 +4,11 @@ import { RealDone } from '@src/components/Metrics/MetricsStep/RealDone'
 import { CycleTime } from '@src/components/Metrics/MetricsStep/CycleTime'
 import { Classification } from '@src/components/Metrics/MetricsStep/Classification'
 import { selectDateRange, selectJiraColumns, selectMetrics, selectUsers } from '@src/context/config/configSlice'
-import { DONE, HEADER_NOTIFICATION_MESSAGE, REQUIRED_DATA } from '@src/constants'
+import { DONE, REQUIRED_DATA } from '@src/constants'
 import { selectCycleTimeSettings, selectMetricsContent } from '@src/context/Metrics/metricsSlice'
 import { DeploymentFrequencySettings } from '@src/components/Metrics/MetricsStep/DeploymentFrequencySettings'
 import CollectionDuration from '@src/components/Common/CollectionDuration'
 import { MetricSelectionWrapper, MetricsSelectionTitle } from '@src/components/Metrics/MetricsStep/style'
-import { useEffect } from 'react'
-import { useNotificationContext } from '@src/hooks/useNotificationContext'
 
 const MetricsStep = () => {
   const requiredData = useAppSelector(selectMetrics)
@@ -19,19 +17,11 @@ const MetricsStep = () => {
   const targetFields = useAppSelector(selectMetricsContent).targetFields
   const cycleTimeSettings = useAppSelector(selectCycleTimeSettings)
   const { startDate, endDate } = useAppSelector(selectDateRange)
-  const { setNotificationProps } = useNotificationContext()
   const isShowCrewsAndRealDone =
     requiredData.includes(REQUIRED_DATA.VELOCITY) ||
     requiredData.includes(REQUIRED_DATA.CYCLE_TIME) ||
     requiredData.includes(REQUIRED_DATA.CLASSIFICATION)
   const isShowRealDone = cycleTimeSettings.every((e) => e.value !== DONE)
-
-  useEffect(() => {
-    setNotificationProps({
-      open: true,
-      title: HEADER_NOTIFICATION_MESSAGE.REPORT.replace('%s', '30'),
-    })
-  }, [])
 
   return (
     <>
